adds users and tasks endpoints

+++ routes/task.js
@@ -1,15 +1,42 @@
 const express = require("express");
 const router = express.Router();
 const client = require("../databasepg");
+const jwt = require("jsonwebtoken");
 
 router.get("/", (req, res, next) => {
   const GetAllTasks = async () => {
     try {
-      const allTasks = await client.query("SELECT * FROM tasks");
+      // verify token start
+      const authtokenwithbearer = req.headers.authorization;
+      if (!authtokenwithbearer) {
+        return res
+          .status(401)
+          .json({ error: "Unauthorized: Token not provided" });
+      }
+      const authtoken = authtokenwithbearer.split(" ")[1];
+      const decoded = await jwt.verify(authtoken, "privatekey");
+      if (!decoded) {
+        return res
+          .status(401)
+          .json({ error: "Unauthorized: Token not provided" });
+      }
+      const user = await client.query(
+        "SELECT * FROM users WHERE user_id = $1",
+        [decoded.user_id]
+      );
+      if (user.rows.length === 0) {
+        return res.status(404).json({ error: "User not found" });
+      }
+      // verify token end
+
+      const allTasks = await client.query(
+        "SELECT * FROM tasks WHERE assigned_by = $1 OR assigned_to = $1",
+        [decoded.user_id]
+      );
       res.json(allTasks.rows);
     } catch (err) {
       console.log(err);
-      throw new Error(err.message);
+      return next(err);
     }
   };
 
@@ -19,7 +46,29 @@ router.get("/", (req, res, next) => {
 });
 
 router.post("/", (req, res, next) => {
-  const AddTask = async () => {
+  const addtask = async () => {
+    // verify token start
+    const authtokenwithbearer = req.headers.authorization;
+    if (!authtokenwithbearer) {
+      return res
+        .status(401)
+        .json({ error: "Unauthorized: Token not provided" });
+    }
+    const authtoken = authtokenwithbearer.split(" ")[1];
+    const decoded = await jwt.verify(authtoken, "privatekey");
+    if (!decoded) {
+      return res
+        .status(401)
+        .json({ error: "Unauthorized: Token not provided" });
+    }
+    const user = await client.query("SELECT * FROM users WHERE user_id = $1", [
+      decoded.user_id,
+    ]);
+    if (user.rows.length === 0) {
+      return res.status(404).json({ error: "User not found" });
+    }
+    // verify token end
+
     const {
       string_id,
       title,
@@ -28,17 +77,27 @@ router.post("/", (req, res, next) => {
       dateofcreation,
       dueDate,
       dateCompleted,
+      assigned_to,
     } = req.body;
 
     if (!title || !dueDate) {
-      req.status(400).json({ message: "Please provide title and due date" });
-      console.log("error");
-      return;
+      return res
+        .status(400)
+        .json({ message: "Please provide title and due date" });
     }
 
+    const assignee = await client.query(
+      "SELECT * FROM users WHERE email = $1",
+      [assigned_to]
+    );
+
+    if (assignee.rows.length === 0) {
+      return res.status(404).json({ error: "Assignee not found" });
+    }
+    const assigneeId = assignee.rows[0].user_id;
     try {
       const newTask = await client.query(
-        "INSERT INTO tasks (string_id, title, completed, description, dateofcreation, dueDate, dateCompleted) VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *",
+        "INSERT INTO tasks (string_id, title, completed, description, dateofcreation, dueDate, dateCompleted, assigned_by, assigned_to) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) RETURNING *",
         [
           string_id,
           title,
@@ -47,17 +106,18 @@ router.post("/", (req, res, next) => {
           dateofcreation,
           dueDate,
           dateCompleted,
+          decoded.user_id,
+          assigneeId,
         ]
       );
 
       res.json(newTask.rows[0]);
     } catch (err) {
       console.log(err);
-      throw new Error(err.message);
+      return next(err);
     }
   };
-
-  AddTask()
+  addtask()
     .then(() => next())
     .catch((err) => next(err));
 });
@@ -66,6 +126,28 @@ router.delete("/:taskId", (req, res, next) => {
   const taskId = req.params.taskId;
   const DeleteTask = async () => {
     try {
+      // verify token start
+      const authtokenwithbearer = req.headers.authorization;
+      if (!authtokenwithbearer) {
+        return res
+          .status(401)
+          .json({ error: "Unauthorized: Token not provided" });
+      }
+      const authtoken = authtokenwithbearer.split(" ")[1];
+      const decoded = await jwt.verify(authtoken, "privatekey");
+      if (!decoded) {
+        return res
+          .status(401)
+          .json({ error: "Unauthorized: Token not provided" });
+      }
+      const user = await client.query(
+        "SELECT * FROM users WHERE user_id = $1",
+        [decoded.user_id]
+      );
+      if (user.rows.length === 0) {
+        return res.status(404).json({ error: "User not found" });
+      }
+      // verify token end
       const result = await client.query(
         "DELETE FROM tasks WHERE string_id = $1",
         [taskId]
@@ -86,6 +168,27 @@ router.delete("/:taskId", (req, res, next) => {
 router.put("/updatecompletestatdb/:taskId", (req, res, next) => {
   const taskId = req.params.taskId;
   const updatecompletionstatus = async () => {
+    // verify token start
+    const authtokenwithbearer = req.headers.authorization;
+    if (!authtokenwithbearer) {
+      return res
+        .status(401)
+        .json({ error: "Unauthorized: Token not provided" });
+    }
+    const authtoken = authtokenwithbearer.split(" ")[1];
+    const decoded = await jwt.verify(authtoken, "privatekey");
+    if (!decoded) {
+      return res
+        .status(401)
+        .json({ error: "Unauthorized: Token not provided" });
+    }
+    const user = await client.query("SELECT * FROM users WHERE user_id = $1", [
+      decoded.user_id,
+    ]);
+    if (user.rows.length === 0) {
+      return res.status(404).json({ error: "User not found" });
+    }
+    // verify token end
     const { completed, dateCompleted } = req.body;
 
     try {
@@ -109,6 +212,27 @@ router.put("/updatecompletestatdb/:taskId", (req, res, next) => {
 router.put("/edit/:taskId", (req, res, next) => {
   const taskId = req.params.taskId;
   const EditTask = async () => {
+    // verify token start
+    const authtokenwithbearer = req.headers.authorization;
+    if (!authtokenwithbearer) {
+      return res
+        .status(401)
+        .json({ error: "Unauthorized: Token not provided" });
+    }
+    const authtoken = authtokenwithbearer.split(" ")[1];
+    const decoded = await jwt.verify(authtoken, "privatekey");
+    if (!decoded) {
+      return res
+        .status(401)
+        .json({ error: "Unauthorized: Token not provided" });
+    }
+    const user = await client.query("SELECT * FROM users WHERE user_id = $1", [
+      decoded.user_id,
+    ]);
+    if (user.rows.length === 0) {
+      return res.status(404).json({ error: "User not found" });
+    }
+    // verify token end
     const { title, description, dueDate } = req.body;
 
     try {
